+++ src/Components/FirstPage/Authentication.module.scss
@@ -1,0 +1,90 @@
+.container_div {
+    height: auto;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: #40BDB6;
+
+
+    .main_div {
+        // padding: 0;
+        min-width: 16.5rem;
+        width: 41.5rem;
+        // height: 100vh;
+
+
+        .middle_div {
+
+            height: 15vh;
+            background-color: white;
+            border-radius: 0% 0% 50% 50%;
+
+        }
+
+        .upper_div {
+            // min-width: 260px;
+            // width: 660px;
+            height: 30vh;
+            background-color: white;
+            display: flex;
+            justify-content: center;
+            align-items: flex-end;
+
+            .logo_div {
+                width: 6.5rem;
+                // height: 50px;
+            }
+        }
+
+        .lower_div {
+            // margin-top: 20px;
+            padding: 3.5rem;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            flex-direction: column;
+            height: 55vh;
+            background-color: #40BDB6;
+            gap: 1.5rem;
+            // position: relative;
+            // overflow: hidden;
+
+            .div_1 {
+                display: flex;
+                flex-direction: column;
+                justify-content: center;
+                align-items: center;
+
+
+
+                button {
+                    width: 3.2rem;
+                    height: 2.5rem;
+                    padding: 0.3rem;
+                    border-radius: 0.5rem;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    border: none;
+                    margin: 0.7rem;
+                }
+
+                .div_2 {
+                    color: #ffffff;
+
+                    p {
+                        font-family: 'Merienda', sans-serif;
+                        font-size: 20px;
+                        font-weight: 400;
+                        line-height: 72px;
+                        letter-spacing: 0em;
+                        text-align: center;
+                        margin-bottom: 0.45rem;
+
+                    }
+                }
+            }
+        }
+    }
+}
+
