--- src/Components/LoginPage/Authentication.scss
@@ -1,92 +1,0 @@
-.container_div {
-    height: auto;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #40BDB6;
-    
-
-    .main_div {
-        // padding: 0;
-        min-width: 260px;
-        width: 660px;
-        // height: 100vh;
-    
-
-    .middle_div {
-    
-        height: 15vh;
-        background-color: white;
-        border-radius: 0% 0% 50% 50%;
-
-    }
-
-    .upper_div {
-        // min-width: 260px;
-        // width: 660px;
-        height: 30vh;
-        background-color: white;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        .logo_div{
-            width: 100px;
-            // height: 50px;
-        }
-    }
-
-    .lower_div {
-        // margin-top: 20px;
-        padding: 55px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-direction: column;
-        height: 55vh;
-        background-color: #40BDB6;
-        gap: 20px;
-        // position: relative;
-        // overflow: hidden;
-
-        .div_1 {
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
-
-
-
-            button {
-                width: 50px;
-                height: 36px;
-                padding: 4px;
-                border-radius: 8px;
-                display: flex;
-                justify-content: center;
-                align-items: center;
-                border: none;
-                margin: 10px;
-            }
-
-            .div_2 {
-                color: #ffffff;
-
-                p{
-                    font-family: 'Merienda',sans-serif;
-                    font-size: 20px;
-                    font-weight: 400;
-                    line-height: 72px;
-                    letter-spacing: 0em;
-                    text-align: center;
-                    margin-bottom: 7px;
-
-                }
-            }
-
-
-        }
-    }
-
-}
-}
-
