+++ src/Components/LoginPage/LoginPage.jsx
@@ -1,52 +1,36 @@
-import React, { useState } from 'react';
-import './Authentication.scss';
-import Lottie from "lottie-react";
-import animationData from "./Comp.json";
-import { useNavigate } from "react-router-dom";
+import React from 'react';
+import styles from './Login.module.scss';
+import { Link } from 'react-router-dom';
 
-
-const LoginPage = () => {
-
-    const navigate = useNavigate();
-
-    const Login = () => {
-        navigate('/login');
-  };
-    
-    
+function LoginPage(props) {
     return (
-
-        <div className='container_div'>
-            <div className='main_div'>
-                <div className='upper_div'>
-                    <div className='logo_div'>
-                        <Lottie animationData={animationData}  loop={false} />
+        <>
+            <div className={styles.container}>
+                <div className={styles.main}>
+                    <div className={styles.upper_div}>
+                        <div className={styles.logo}>
+                        <img src="https://res.cloudinary.com/dlx4meooj/image/upload/c_pad,b_auto:predominant,fl_preserve_transparency/v1694542881/Frame_58099_igiaij.jpg?_s=public-apps" alt="" />
+                        </div>
                     </div>
-                </div>
-
-                <div className='middle_div'></div>
-                <div className='lower_div'>
-                    <div className='div_1'>
-                        
-                            <button className='student_div' onClick={Login}>
-
-                                <img src="https://res.cloudinary.com/dlx4meooj/image/upload/c_pad,b_auto:predominant,fl_preserve_transparency/v1694618443/Property_1_Default_3_1_p791jw.jpg?_s=public-apps" alt="" />
-
-                            </button>
-                        
-                        
-                            <button className='admin_div' onClick={Login}>
-
-                                <img src="https://res.cloudinary.com/dlx4meooj/image/upload/c_pad,b_auto:predominant,fl_preserve_transparency/v1694618443/Property_1_Default_1_w8h3pa.jpg?_s=public-apps" alt="" />
-                            </button>
-                       
-                    </div>
-                    <div className='div_2'>
-                        <p>Vyatha</p>
+                {/* <p>Login as {props.type=="student"?"Student":"Admin"}</p> */}
+                    <div className={styles.lower_div}>
+                        <div className={styles.content_div}>
+                            <div className={styles.image}>
+                                <button className={styles.login}>
+                                    <img src="https://res.cloudinary.com/dlx4meooj/image/upload/c_pad,b_auto:predominant,fl_preserve_transparency/v1695295732/login_w7ltqc.jpg?_s=public-apps" alt="" />
+                                </button>
+                            </div>
+                            <div className={styles.content}>
+                                <p>Don't have an account? <Link to={''}>Sign up</Link></p>
+                            </div>
+                        </div>
+                        <div className={styles.text}>
+                            <p>Vyatha</p>
+                        </div>
                     </div>
                 </div>
             </div>
-        </div>
+        </>
     )
 }
